% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cellustering.R
\name{Cellustering}
\alias{Cellustering}
\title{Initialize a Cellustering instance}
\usage{
Cellustering(data)
}
\arguments{
\item{data}{The data storing 10X single-cell RNA data}
}
\value{
A fresh Cellustering instance with data describing the cell names
and their features.
}
\description{
Initialize a Cellustering instance
}
