% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{cellustering_example}
\alias{cellustering_example}
\title{Get path to Cellustering example}
\usage{
cellustering_example(file = NULL)
}
\arguments{
\item{file}{Name of file. If \code{NULL}, the example files will be listed.}
}
\description{
cellustering comes bundled with a number of sample files in its
\code{inst/extdata} directory. This function make them easy to access
}
\examples{
cellustering_example()
cellustering_example("hg19")
}
