% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normalize.R
\name{normalize}
\alias{normalize}
\title{Function Description}
\usage{
normalize(object, scale_factor = 1e+06, log_transformation = TRUE)
}
\arguments{
\item{object}{Parameter one Description.}

\item{scale_factor}{Parameter two Description.}
}
\value{
Parameter two Description.
}
\description{
Function Description
}
\examples{
normalize(123) # report error
normalize(pbmc, scale.factor = -100) # report error
normalize(pbmc, log.transformation = "abc") # report error
find_HVG(pbmc) # report error
col_sum_before <- sum(pbmc@data[, 1])
element_before <- pbmc@data[30, 1]
pbmc <- normalize(pbmc, scale_factor = 1e6)
col_sum_after <- sum(pbmc@data[, 1])
element_after <- pbmc@data[30, 1]
all.equal(element.after, log(element.before / col.sum.before * 1e6 + 1))
all.equal(sum(exp(pbmc@data[, 1]) - 1), 1e6)
}
