% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{pbmc_small}
\alias{pbmc_small}
\title{PBMC Small}
\format{
\code{ pbmc_small}
A \code{Cellustering} instance whose data is a dataframe with 230 rows and 80
columns:
}
\source{
\url{https://github.com/satijalab/seurat}
}
\usage{
pbmc_small
}
\description{
The \code{pbmc_small} is a tiny dataset tailored from the \code{pbmc} dataset for
illustration the entire workflow of the \code{cellustering} package.
}
\keyword{datasets}
