% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dimension_deduction.R
\name{principal_component_analysis}
\alias{principal_component_analysis}
\title{Perform Principal Component Analysis (PCA) and visualize the data with two
specified principal components (PCs).}
\usage{
principal_component_analysis(object, PC1 = 1, PC2 = 2)
}
\arguments{
\item{object}{A \code{Cellustering} instance.}

\item{PC1}{The index of the first principal component to be visualized.}

\item{PC2}{The index of the second principal component to be visualized.}
}
\value{
The \code{Cellustering} instance with singular value decomposition (SVD)
information and PCA plot added to the \code{reduced_dimension} slot.
}
\description{
Perform Principal Component Analysis (PCA) and visualize the data with two
specified principal components (PCs).
}
\examples{
# principal_component_analysis(123) # report error
# principal_component_analysis(pbmc, PC1 = 0.1, PC2 = 3000)  # report error

# Run the necessary to catch up the progress
data("pbmc_small")
pbmc_small_7 <- Cellustering(pbmc_small@data)
pbmc_small_7 <- qc_plot(pbmc_small_7)
pbmc_small_7 <- qc_filter(pbmc_small_7)
pbmc_small_7 <- normalize(pbmc_small_7, scale_factor = 1e6)
pbmc_small_7 <- find_HVG(pbmc_small_7, n_feature = 20)
pbmc_small_7 <- scale_data(pbmc_small_7)

pbmc_small_7 <- principal_component_analysis(pbmc_small_7)

}
