% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocessing.R
\name{qc_filter}
\alias{qc_filter}
\title{Filter out low-quality cells and genes based on user-defined conditions.}
\usage{
qc_filter(
  object,
  min_count_depth = 0,
  max_count_depth = Inf,
  min_feature = 0,
  max_feature = Inf,
  max_mito_percent = 100,
  min_total_expression = 0,
  min_cell = 0
)
}
\arguments{
\item{object}{A \code{Cellustering} instance.}

\item{min_count_depth}{The minimum count depth allowed for a cell.}

\item{max_count_depth}{The maximum count depth allowed for a cell.}

\item{min_feature}{The minimum number of features allowed for a cell.}

\item{max_feature}{The maximum number of features allowed for a cell.}

\item{max_mito_percent}{The maximum mitochondrial gene percentage allowed
for a cell.}

\item{min_total_expression}{The minimum total expression allowed for a
gene.}

\item{min_cell}{The minimum number of cells in which a gene must be
expressed.}
}
\value{
A \code{Cellustering} instance with cells and genes satisfying the
filtering conditions.
}
\description{
Filter out low-quality cells and genes based on user-defined conditions.
}
\examples{
# qc_filter(123) # report Error
# qc_filter(pbmc, min.count.depth = -1) # report Error
# qc_filter(pbmc, min.cells = "abc") # report Error
# Run the necessary to catch up the progress
pbmc_small <- qc_plot(pbmc_small)

pbmc_small <- qc_filter(pbmc_small)
dim(pbmc_small@data)

pbmc_small <- qc_filter(pbmc_small,
  min_feature = 200, max_feature = 2500,
  max_mito_percent = 5, min_cell = 3
)
dim(pbmc_small@data)
# Then we check the plots again
pbmc_small <- qc_plot(pbmc_small)
}
