% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocessing.R
\name{qc_plot}
\alias{qc_plot}
\title{Function to plot distribution of total count and feature count or violin plot}
\usage{
qc_plot(object)
}
\arguments{
\item{object}{A \code{Cellustering} instance.}
}
\value{
A \code{Cellusetering} instance that ready to generate plot and for later
process.
}
\description{
Function to plot distribution of total count and feature count or violin plot
}
\examples{
pbmc <- read_10x("hg19")
# qc_plot(123) # report Error
# qc_filter(pbmc) # report Error
pbmc_small <- qc_plot(pbmc_small)
head(pbmc@quality$cell_data)
head(pbmc@quality$feature_data)
}
