% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/initialize.R
\name{read_10x}
\alias{read_10x}
\title{Initialize a \code{Cellustering} instance from a directory containing
files storing single cell RNA data in the 10x standard.}
\usage{
read_10x(directory)
}
\arguments{
\item{directory}{A directory containing single cell RNA data.}
}
\value{
A \code{Cellustering} instance.
}
\description{
Initialize a \code{Cellustering} instance from a directory containing
files storing single cell RNA data in the 10x standard.
}
\examples{
pbmc <- read_10x("hg19")
pbmc@data[c("CD3D", "TCL1A", "MS4A1"), 1:30]
}
