% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scale_data.R
\name{scale_data}
\alias{scale_data}
\title{Scale and center the selected features (HVG) data，ensuring that each HVG's
expression counts have mean 0 and variance 1.}
\usage{
scale_data(object)
}
\arguments{
\item{object}{A \code{Cellustering} instance.}
}
\value{
The \code{Cellusetering} instance with the scaled data added to the
\code{reduced_dimension} slot.
}
\description{
Scale and center the selected features (HVG) data，ensuring that each HVG's
expression counts have mean 0 and variance 1.
}
\examples{
# Run the necessary to catch up the progress
data("pbmc_small")
pbmc_small_6 <- Cellustering(pbmc_small@data)
pbmc_small_6 <- qc_plot(pbmc_small_6)
pbmc_small_6 <- qc_filter(pbmc_small_6)
pbmc_small_6 <- normalize(pbmc_small_6, scale_factor = 1e6)
pbmc_small_6 <- find_HVG(pbmc_small_6, n_feature = 20)

pbmc_small_6 <- scale_data(pbmc_small_6)
pbmc_small_6@reduced_dimension$scaled_data[1:5, 1:10]

# Check if the row sums equal to 0
row_sums <- rowSums(pbmc_small_6@reduced_dimension$scaled_data)
all.equal(row_sums, rep(0, 2000), check.attributes = FALSE)

# Check if the row sd equal to 1
sds <- apply(pbmc_small_6@reduced_dimension$scaled_data, 1, sd)
all.equal(sds, rep(1, 2000), check.attributes = FALSE)
}
