% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dimension_deduction.R
\name{select_proper_dimension}
\alias{select_proper_dimension}
\title{Function Description}
\usage{
select_proper_dimension(object, lower_bound = 1, upper_bound = 30)
}
\arguments{
\item{parameter_1}{Parameter one Description.}

\item{parameter_2}{Parameter two Description.}
}
\value{
Parameter two Description.
}
\description{
Function Description
}
\examples{
 x <- test_value
 function(test_value)
}
