% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cellustering.R
\name{Cellustering}
\alias{Cellustering}
\title{Initialize a Cellustering instance}
\usage{
Cellustering(
  data,
  quality = list(),
  HVG = list(),
  reduced_dimension = list(),
  clustering = list()
)
}
\arguments{
\item{parameter_1}{Parameter one Description.}

\item{parameter_2}{Parameter two Description.}
}
\value{
Parameter two Description.
}
\description{
Initialize a Cellustering instance
}
