% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clustering.R
\name{compute_silhouette}
\alias{compute_silhouette}
\title{Calculate the average silhouette coefficient for the clustering results.}
\usage{
compute_silhouette(object)
}
\arguments{
\item{object}{A \code{Cellustering} instance.}
}
\value{
The \code{Cellustering} instance with the average silhouette coefficient
added to the \code{clustering} slot.
}
\description{
Calculate the average silhouette coefficient for the clustering results.
}
\examples{
# Run the necessary to catch up the progress
data("pbmc_small")
pbmc_small_0 <- Cellustering(pbmc_small@data)
pbmc_small_0 <- qc_plot(pbmc_small_0)
pbmc_small_0 <- qc_filter(pbmc_small_0)
pbmc_small_0 <- normalize(pbmc_small_0, scale_factor = 1e6)
pbmc_small_0 <- find_HVG(pbmc_small_0, n_feature = 20)
pbmc_small_0 <- scale_data(pbmc_small_0)
pbmc_small_0 <- principal_component_analysis(pbmc_small_0)
pbmc_small_0 <- select_proper_dimension(pbmc_small_0,
  lower_bound = 1, upper_bound = 5
)
pbmc_small_0 <- kmeans(pbmc_small_0, 2)

pbmc_small_0 <- compute_silhouette(pbmc_small_0)
}
