% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/feature_selection.R
\name{find_HVG}
\alias{find_HVG}
\title{Identify, visualize, and select features called highly variable genes (HVGs).}
\usage{
find_HVG(object, n_feature = 2000, loess_span = 0.5)
}
\arguments{
\item{object}{A \code{Cellustering} instance.}

\item{n_feature}{The number of features to select.}

\item{loess_span}{The span parameter for the local regression model,
controlling smoothness of the fitted curve.}
}
\value{
The \code{Cellusetering} instance with HVG names and plot added to the
\code{HVG} slot.
}
\description{
Identify, visualize, and select features called highly variable genes (HVGs).
}
\examples{
# find_HVG(123) # report error
# scale_data(pbmc) # report error

# Run the necessary to catch up the progress
data("pbmc_small")
pbmc_small_5 <- Cellustering(pbmc_small@data)
pbmc_small_5 <- qc_plot(pbmc_small_5)
pbmc_small_5 <- qc_filter(pbmc_small_5)
pbmc_small_5 <- normalize(pbmc_small_5, scale_factor = 1e6)

pbmc_small_5 <- find_HVG(pbmc_small_5, n_feature = 20)
pbmc_small_5@HVG$HVG_plot
}
