% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/feature_selection.R
\name{find_HVG}
\alias{find_HVG}
\title{Function Description}
\usage{
find_HVG(object, n_feature = 2000, loess_span = 0.5)
}
\arguments{
\item{parameter_1}{Parameter one Description.}

\item{parameter_2}{Parameter two Description.}
}
\value{
Parameter two Description.
}
\description{
Function Description
}
\examples{
# find_HVG(123) # report error
# scale_data(pbmc) # report error
pbmc <- find_HVG(pbmc)
pbmc@HVG$HVG_plot
}
