% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clustering.R
\name{kmeans}
\alias{kmeans}
\title{Function Description}
\usage{
kmeans(object, k, dimensions = 10, umap = FALSE)
}
\arguments{
\item{parameter_1}{Parameter one Description.}

\item{parameter_2}{Parameter two Description.}
}
\value{
Parameter two Description.
}
\description{
Function Description
}
\examples{
 x <- test_value
 function(test_value)
}
