% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normalize.R
\name{normalize}
\alias{normalize}
\title{Normalize the data using the Counts per Million (CPM) normalization
protocol. The purpose of normalization is to remove the variations in count
depths among different cells, and to place the gene expression levels of
different cells on a comparable scale.}
\usage{
normalize(object, scale_factor = 1e+06, log_transformation = TRUE)
}
\arguments{
\item{object}{A \code{Cellustering} instance.}

\item{scale_factor}{The level to which each cell's total count depth is
normalized.}

\item{log_transformation}{Determine whether to perform log-transformation
after the CPM normalization.}
}
\value{
The \code{Cellustering} instance with normalized data stored in the
\code{data} slot.
}
\description{
Normalize the data using the Counts per Million (CPM) normalization
protocol. The purpose of normalization is to remove the variations in count
depths among different cells, and to place the gene expression levels of
different cells on a comparable scale.
}
\examples{
# normalize(123) # report error
# normalize(pbmc, scale.factor = -100) # report error
# normalize(pbmc, log.transformation = "abc") # report error
# find_HVG(pbmc) # report error

# Run the necessary to catch up the progress
data("pbmc_small")
pbmc_small_4 <- Cellustering(pbmc_small@data)
pbmc_small_4 <- qc_plot(pbmc_small_4)
pbmc_small_4 <- qc_filter(pbmc_small_4)

col_sum_before <- sum(pbmc_small_4@data[, 1])
element_before <- pbmc_small_4@data[30, 1]
pbmc_small_4 <- normalize(pbmc_small_4, scale_factor = 1e6)
col_sum_after <- sum(pbmc_small_4@data[, 1])
element_after <- pbmc_small_4@data[30, 1]
all.equal(element_after, log(element_before / col_sum_before * 1e6 + 1))
all.equal(sum(exp(pbmc_small_4@data[, 1]) - 1), 1e6)
}
