% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dimension_deduction.R
\name{principal_component_analysis}
\alias{principal_component_analysis}
\title{Function Description}
\usage{
principal_component_analysis(object, PC1 = 1, PC2 = 2)
}
\arguments{
\item{object}{Parameter one Description.}
}
\value{
Parameter two Description.
}
\description{
Function Description
}
\examples{
# principal_component_analysis(123) # report error
# principal_component_analysis(pbmc, PC1 = 0.1, PC2 = 3000)  # report error
pbmc <- principal_component_analysis(pbmc, PC1 = 3, PC2 = 4)
}
