% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocessing.R
\name{qc_filter}
\alias{qc_filter}
\title{Function Description}
\usage{
qc_filter(
  object,
  min_count_depth = 0,
  max_count_depth = Inf,
  min_feature = 0,
  max_feature = Inf,
  max_mito_percent = 100,
  min_total_expression = 0,
  min_cell = 0
)
}
\arguments{
\item{parameter_1}{Parameter one Description.}

\item{parameter_2}{Parameter two Description.}
}
\value{
Parameter two Description.
}
\description{
Function Description
}
\examples{
# qc_filter(123) # report Error
# qc_filter(pbmc, min.count.depth = -1) # report Error
# qc_filter(pbmc, min.cells = "abc") # report Error
pbmc <- read_10x("hg19")
pbmc <- qc_filter(pbmc)
dim(pbmc@data)

pbmc <- qc_filter(pbmc,
  min_feature = 200, max_feature = 2500,
  max_mito_percent = 5, min_cell = 3
)
dim(pbmc@data)
# Then we check the plots again
pbmc <- qc_plot(pbmc)
head(pbmc@QC$Cell.data)
head(pbmc@QC$Feature.data)
pbmc@QC$Cell.plot
pbmc@QC$Feature.plot
pbmc@QC$Violin.plot
}
