% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scale_data.R
\name{scale_data}
\alias{scale_data}
\title{Function Description}
\usage{
scale_data(object)
}
\arguments{
\item{object}{Parameter one Description.}
}
\value{
Parameter two Description.
}
\description{
Function Description
}
\examples{
pbmc <- scale_data(pbmc)
pbmc@reduced_dimension$scaled_data[1:5, 1:10]

# Check if the row sums equal to 0
row.sums <- rowSums(pbmc@reduced_dimension$scaled_data)
all.equal(row.sums, rep(0, 2000), check.attributes = FALSE)

# Check if the row sd equal to 1
sds <- apply(pbmc@reduced_dimension$scaled_data, 1, sd)
all.equal(sds, rep(1, 2000), check.attributes = FALSE)
}
