% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dimension_deduction.R
\name{select_proper_dimension}
\alias{select_proper_dimension}
\title{Determine an appropriate dimension for data to be reduced to.}
\usage{
select_proper_dimension(object, lower_bound = 1, upper_bound = 30)
}
\arguments{
\item{object}{A \code{Cellustering} instance.}

\item{lower_bound}{The minimum number of principal components to consider.}

\item{upper_bound}{The maximum number of principal components to consider.}
}
\value{
The \code{Cellusetering} instance with suggested dimension and plots
added to the \code{reduced_dimension} slot.
}
\description{
Determine an appropriate dimension for data to be reduced to.
}
\examples{
# Run the necessary to catch up the progress
data("pbmc_small")
pbmc_small_8 <- Cellustering(pbmc_small@data)
pbmc_small_8 <- qc_plot(pbmc_small_8)
pbmc_small_8 <- qc_filter(pbmc_small_8)
pbmc_small_8 <- normalize(pbmc_small_8, scale_factor = 1e6)
pbmc_small_8 <- find_HVG(pbmc_small_8, n_feature = 20)
pbmc_small_8 <- scale_data(pbmc_small_8)
pbmc_small_8 <- principal_component_analysis(pbmc_small_8)

pbmc_small_8 <- select_proper_dimension(pbmc_small_8,
  lower_bound = 1, upper_bound = 5
)
}
